refactor : api 폴더화 및 NotFound 들여쓰기 수정 및 BlogComment -> PostComment로 명칭 수정
