+++ 신동현/src/api/login/login.ts
@@ -1,4 +1,4 @@
-import { BaseUrl, api } from "./api";
+import { BaseUrl, api } from "../api";
 
 const EamilLogin = async (email: string, password: string) => {
   try {
