+++ 신동현/src/pages/Error/NotFound.tsx
@@ -1,9 +1,9 @@
 const NotFound = () => {
-    return (
-        <div>
-            <h1>404 Not Found.</h1>
-        </div>
-    )
+  return (
+    <div>
+      <h1>404 Not Found.</h1>
+    </div>
+  )
 }
 
 export default NotFound;
