+++ 신동현/src/pages/index.tsx
@@ -2,7 +2,7 @@ import Home from "./Home";
 import SignUp from "./SignUp/SignUp";
 import SignUpDetail from "./SignUp/SignUpDetail";
 import EmailForm from "./SignUp/EmailForm"; 
-import KaKaoForm from "./SignUp/KaKaoForm";
+import KaKaoForm from "./SignUp/KakaoForm";
 import Mypage from "./Mypage";
 import KakaoLoginLoading from "./SignUp/KakaoLoginLoading";
 import Detail from "./Detail/Detail";
